Fix dottest

+++ aljabr/linop.py
@@ -824,8 +824,8 @@ def dottest(
     """
     test = True
     for _ in range(num):
-        vvec = randn(linop.isize).astype(linop.dtype)
-        uvec = randn(linop.osize).astype(linop.dtype)
+        vvec = randn(linop.isize)
+        uvec = randn(linop.osize)
         test = test & np.allclose(
             i := np.dot(linop.rmatvec(uvec).conj().ravel(), vvec.ravel()),
             j := np.dot(uvec.conj().ravel(), linop.matvec(vvec).ravel()),
